simg2img.js script
- Replace "sparse benchmark" with unsparsify script
- Use in CI
- Replace "compare output" test in sparse.spec.js with simg2img.js and
  cmp with Bun shell
